refactor: restructure imports to follow Python best practices
- Replace magic __init__.py exports with explicit module imports
- Add src directory to Python path in main.py for proper absolute imports  
- Change from 'import util' to specific imports: 'from util.replicate_api import Replicate_API'
- Standardize all GUI modules to use 'from util.settings import Settings'
- Remove __init__.py re-exports to eliminate import magic
- Enable running application from project root with 'python3 src/main.py'
- Improve IDE support and import clarity for better maintainability
