+++ src/gui/__init__.py
@@ -1,6 +1,1 @@
-from .filehandler import FileHandler
-from .imagegenerator import ImageGeneratorGUI
-from .lightbox import Lightbox
-from .styles import Styles
-from .usermodels import UserModels
