--- src/util/__init__.py
@@ -1,3 +1,0 @@
-# from .settings import Settings
-from .replicate_api import Replicate_API
-from .settings import Settings
